🐛 WIP, analogio - Refactor hardware.cpp for raw adc reads, fix underflow bug in func signature mismatch for model

+++ src/components/analogIO/model.cpp
@@ -146,8 +146,8 @@ bool AnalogIOModel::EncodeAnalogIOEvent(
     @return True if successful, False otherwise.
 */
 /********************************************************************************/
-bool AnalogIOModel::EncodeAnalogIOEventRaw(char *pin_name, int16_t pin_value) {
-  return EncodeAnalogIOEvent(pin_name, (float)pin_value,
+bool AnalogIOModel::EncodeAnalogIOEventRaw(char *pin_name, float pin_value) {
+  return EncodeAnalogIOEvent(pin_name, pin_value,
                              wippersnapper_sensor_SensorType_SENSOR_TYPE_RAW);
 }
